put shared types in a separate file
removes ts compiler circular dependency warning
